@@ -4,7 +4,16 @@ import { Routes, RouterModule } from '@angular/router';
 const routes: Routes = [
   {
     path: '',
-    loadChildren: './file-explorer/file-explorer.module#FileExplorerModule'
+    pathMatch: 'full',
+    redirectTo: 'explorer',
+  },
+  {
+    path: 'passport',
+    loadChildren: () => import('./passport/passport.module').then(module => module.PassportModule),
+  },
+  {
+    path: 'explorer',
+    loadChildren: () => import('./file-explorer/file-explorer.module').then(module => module.FileExplorerModule),
   },
 ];
 
