@@ -4,6 +4,8 @@ import { ListsModule } from './lists/public_api';
 import { CommonModule } from '@angular/common';
 import { ScrimsModule } from './scrims/public_api';
 import { ThemesModule } from './themes/public_api';
+import { OthersModule } from './others/public_api';
+import { InputsModule } from './inputs/public_api';
 import { AppBarsModule } from './app-bars/public_api';
 import { ButtonsModule } from './buttons/public_api';
 import { DialogsModule } from './dialogs/public_api';
@@ -12,7 +14,6 @@ import { LoadingsModule } from './loadings/public_api';
 import { CheckboxsModule } from './checkboxs/public_api';
 import { ToolbarComponent } from './toolbar/toolbar.component';
 import { ContentComponent } from './content/content.component';
-import { OthersModule } from './others/public_api';
 
 @NgModule({
   imports: [
@@ -22,6 +23,7 @@ import { OthersModule } from './others/public_api';
     ScrimsModule,
     ThemesModule,
     CommonModule,
+    InputsModule,
     AppBarsModule,
     ButtonsModule,
     DialogsModule,
@@ -35,6 +37,7 @@ import { OthersModule } from './others/public_api';
     OthersModule,
     ScrimsModule,
     ThemesModule,
+    InputsModule,
     AppBarsModule,
     ButtonsModule,
     DialogsModule,
